Move `region_scale` widget to an addon; it was primarily only useful for debugging
